Fixed UTC and no dropdown value issue
When no dropdown was selected it wouldn't have a standard value. This was caused because the standard value was the string name of an enum while it should have been an integer. The UTC issue was caused by the way the string was formatted.

Also I have added the functionality that you don't have to select a time to plan a task for that day.