Tasks Plan display & Bug Fix with Task Day Display
Tasks that don't have a specified time will now show no time.

Tasks who were planned at the start of the month showed up at the end of the month, because IsTaskOnDay only checked the time of current month and year, but didn't take caution with previous and next months. This meant that some tasks were duplicated. With a new method who uses a formula to calculate if it is current, prev or next month has fixed this issue.