Admin Functionalities Added in Vue, Small Refactor in API
Admin Panel is added in Vue and works, except the jwt token doesn't work properly yet. This is due to fact that the Api class only checks once what the token is. Which means that the data of the old token is used even if there is no token when removed.